Refactor ApiV1Dot1Controller into separate organized controllers
- Split ApiV1Dot1Controller into 8 focused controllers:
  - AccountController: Account management, settings, and user data
  - AuthController: Authentication and registration
  - StatusController: Status/post creation
  - ArchiveController: Post archiving functionality
  - PlacesController: Location-based posts
  - ReportController: Content reporting
  - ModerationController: Admin moderation actions
  - PushNotificationController: Push notification management (includes nagState)

- Updated all v1.1 API routes to use new controller structure
- Maintained all existing functionality and method signatures
- Improved code organization and maintainability
- Only affects V1.1 API controllers, V2 controllers remain unchanged
